+++ camera.lua
@@ -24,8 +24,8 @@ function normalCamera(mousex,mousey,dt,rxy)
     
     local remcx = CameraX
     local remcy = CameraY
-    CameraX = CameraX + (tx-CameraX) * 7.5*dt + rxy*GameScale*(love.math.random()-0.5)
-    CameraY = CameraY + (ty-CameraY) * 7.5*dt + rxy*GameScale*(love.math.random()-0.5)
+    CameraX = CameraX + (tx-CameraX) * 7.5*dt + (0.007/dt)*rxy*GameScale*(love.math.random()-0.5)
+    CameraY = CameraY + (ty-CameraY) * 7.5*dt + (0.007/dt)*rxy*GameScale*(love.math.random()-0.5)
     DiffCX = CameraX-remcx
     DiffCY = CameraY-remcy